Backend/requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
pdfplumber==0.11.0
requests==2.32.3
werkzeug==3.0.1
pymongo==4.7.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bson==0.5.10